fixing flatpak pinta with -y

+++ uninstall/app-pinta.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
-
-sudo snap remove pinta
+source ~/.local/share/omakub/install/lib/env.sh
+if [ "$OS_NAME" = "Ubuntu" ]; then
+  sudo snap remove pinta
+else
+  flatpak uninstall -y flathub com.github.PintaProject.Pinta
+fi
